+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - `vagrant ssh` now works on Solaris, where `IdentitiesOnly` was not
     an available option. [GH-820]
   - Output works properly in the face of broken pipes. [GH-819]
+  - Enable Host IO Cache on the SATA controller by default.
 
 ## 1.0.1 (March 11, 2012)